9810 - Fixing import issue + unit test
