+++ web-client/src/views/RecentFilings/RecentFilingsNonMobile.tsx
@@ -14,7 +14,7 @@ import { DOCKET_ENTRY_SEALED_TO_TYPES } from '@shared/business/entities/EntityCo
 import { BigHeader } from '../BigHeader';
 import { RecentFilingsDocumentDisplay } from './RecentFilingsDocumentDisplay';
 import { Button } from '@web-client/ustc-ui/Button/Button';
-import { formatCaseStatus } from '@web-client/presenter/computeds/formattedWorkQueue';
+import { Case } from 'shared/src/business/entities/cases/Case';
 
 type SortableField = 'docketNumber' | 'filedDate' | 'document' | 'caseTitle';
 type SortOrder = 'asc' | 'desc';
@@ -261,7 +261,7 @@ export const RecentFilingsNonMobile = ({
                             })
                           }
                         >
-                          {formatCaseStatus({
+                          {Case.formatCaseStatus({
                             caseStatus: filing.status,
                             trialDate: filing.trialDate,
                             trialLocation: filing.trialLocation,
